+++ src/exactly_lib/instructions/multi_phase/env.py
@@ -5,18 +5,19 @@ from exactly_lib.common.help.instruction_documentation_with_text_parser import \
     InstructionDocumentationThatIsNotMeantToBeAnAssertionInAssertPhaseBase
 from exactly_lib.common.help.syntax_contents_structure import InvokationVariant
 from exactly_lib.definitions import instruction_arguments
-from exactly_lib.definitions.entity.concepts import SYMBOL_CONCEPT_INFO
-from exactly_lib.definitions.entity.syntax_elements import SYMBOL_REFERENCE_SYNTAX_ELEMENT
+from exactly_lib.definitions.entity import concepts
+from exactly_lib.definitions.entity import syntax_elements
 from exactly_lib.instructions.multi_phase.utils import instruction_embryo as embryo
 from exactly_lib.instructions.multi_phase.utils.assert_phase_info import IsAHelperIfInAssertPhase
 from exactly_lib.instructions.multi_phase.utils.instruction_part_utils import PartsParserFromEmbryoParser, \
     MainStepResultTranslatorForUnconditionalSuccess
+from exactly_lib.section_document.element_parsers import token_stream_parser
 from exactly_lib.section_document.element_parsers.instruction_parser_exceptions import \
     SingleInstructionInvalidArgumentException
-from exactly_lib.section_document.element_parsers.misc_utils import new_token_stream, \
+from exactly_lib.section_document.element_parsers.misc_utils import \
     std_error_message_text_for_token_syntax_error_from_exception
-from exactly_lib.section_document.element_parsers.token_stream import TokenStream, TokenSyntaxError
-from exactly_lib.symbol.data.restrictions.reference_restrictions import is_any_data_type
+from exactly_lib.section_document.element_parsers.token_stream import TokenSyntaxError
+from exactly_lib.section_document.element_parsers.token_stream_parser import TokenParser
 from exactly_lib.symbol.data.string_sdv import StringSdv
 from exactly_lib.symbol.path_resolving_environment import PathResolvingEnvironmentPreOrPostSds
 from exactly_lib.symbol.symbol_usage import SymbolUsage, SymbolReference
@@ -24,8 +25,11 @@ from exactly_lib.test_case.os_services import OsServices
 from exactly_lib.test_case.phases.common import InstructionEnvironmentForPostSdsStep, \
     PhaseLoggingPaths
 from exactly_lib.test_case_utils.parse import parse_string
+from exactly_lib.util.string import StringFormatter
 from exactly_lib.util.textformat.structure.core import ParagraphItem
-from exactly_lib.util.textformat.structure.structures import paras
+
+UNSET_IDENTIFIER = 'unset'
+ASSIGNMENT_IDENTIFIER = instruction_arguments.ASSIGNMENT_OPERATOR
 
 
 class TheInstructionDocumentation(InstructionDocumentationThatIsNotMeantToBeAnAssertionInAssertPhaseBase,
@@ -42,39 +46,17 @@ class TheInstructionDocumentation(InstructionDocumentationThatIsNotMeantToBeAnAs
     def invokation_variants(self) -> Sequence[InvokationVariant]:
         return [
             InvokationVariant(
-                _format('{NAME} = {VALUE}'),
+                self._tp.format('{NAME} = {VALUE}'),
                 self._tp.fnap(_DESCRIPTION_OF_SET)),
             InvokationVariant(
-                _format('unset {NAME}'),
-                paras(_format('Removes the environment variable {NAME}.'))),
+                self._tp.format('{unset_keyword} {NAME}'),
+                self._tp.fnap('Removes the environment variable {NAME}.')),
         ]
 
     def see_also_targets(self) -> list:
         return [
-            SYMBOL_REFERENCE_SYNTAX_ELEMENT.cross_reference_target
+            syntax_elements.STRING_SYNTAX_ELEMENT.cross_reference_target,
         ]
-
-
-def _format(template: str) -> str:
-    return template.format_map(_FORMAT_DICT)
-
-
-_FORMAT_DICT = {
-    'NAME': 'NAME',
-    'VALUE': 'VALUE',
-    'SYMBOLS': SYMBOL_CONCEPT_INFO.name.plural,
-}
-
-_DESCRIPTION_OF_SET = """\
-Sets the environment variable {NAME} to {VALUE}.
-
-
-Elements of the form "${{var_name}}" in {VALUE}, will be replaced with the value of the environment variable "var_name",
-or the empty string, if there is no environment variable with that name.
-
-
-{VALUE} may contain references to {SYMBOLS}.
-"""
 
 
 class Executor:
@@ -104,53 +86,50 @@ class TheInstructionEmbryo(embryo.InstructionEmbryo[None]):
                               environment.path_resolving_environment_pre_or_post_sds)
 
 
-class EmbryoParser(embryo.InstructionEmbryoParserThatConsumesCurrentLine):
-    def _parse(self, rest_of_line: str) -> TheInstructionEmbryo:
+class EmbryoParser(embryo.InstructionEmbryoParserFromTokensWoFileSystemLocationInfo[None]):
+    def _parse_from_tokens(self, token_parser: TokenParser) -> TheInstructionEmbryo:
         try:
-            tokens = new_token_stream(rest_of_line)
-            if tokens.is_null:
-                raise SingleInstructionInvalidArgumentException('Missing arguments')
-            first = tokens.consume()
-            if first.is_quoted:
-                raise SingleInstructionInvalidArgumentException('Variable name must not be quoted')
-            if tokens.is_null:
-                if first.string == UNSET_IDENTIFIER:
-                    raise SingleInstructionInvalidArgumentException('Missing variable name to unset')
-                else:
-                    raise SingleInstructionInvalidArgumentException('Missing arguments of variable assignment')
-            second = tokens.consume()
-            if second.source_string == ASSIGNMENT_IDENTIFIER:
-                return self._parse_set(first.string, tokens)
-            elif first.string == UNSET_IDENTIFIER:
-                if second.is_quoted:
-                    raise SingleInstructionInvalidArgumentException('Variable name must not be quoted')
-                return self._parse_unset(second.string, tokens)
+            unset_keyword_or_var_name = token_parser.consume_mandatory_unquoted_string__w_err_msg(
+                False,
+                _MISSING_UNSET_KEYWORD_OR_VAR_NAME_ERROR_MESSAGE
+            )
 
-            else:
-                raise SingleInstructionInvalidArgumentException('Invalid syntax')
+            return (
+                self._parse_unset_or_set_var_with_same_name_as_unset_keyword(token_parser)
+                if unset_keyword_or_var_name == UNSET_IDENTIFIER
+                else
+                self._parse_set(token_parser, unset_keyword_or_var_name)
+            )
+
         except TokenSyntaxError as ex:
             raise SingleInstructionInvalidArgumentException(
                 std_error_message_text_for_token_syntax_error_from_exception(ex))
 
-    def _parse_unset(self, variable_name: str, remaining_tokens: TokenStream) -> TheInstructionEmbryo:
-        if not remaining_tokens.is_null:
-            raise SingleInstructionInvalidArgumentException(_format('Superfluous arguments.'))
-        return TheInstructionEmbryo(_UnsetExecutor(variable_name), [])
+    def _parse_unset_or_set_var_with_same_name_as_unset_keyword(self,
+                                                                token_parser: TokenParser,
+                                                                ) -> TheInstructionEmbryo:
+        if token_parser.has_valid_head_token():
+            head = token_parser.head
+            if head.is_plain and head.string == ASSIGNMENT_IDENTIFIER:
+                return self._parse_set(token_parser, UNSET_IDENTIFIER)
 
-    def _parse_set(self, variable_name: str, tokens_for_value: TokenStream) -> TheInstructionEmbryo:
-        if tokens_for_value.is_null:
-            raise SingleInstructionInvalidArgumentException(_format('Missing {VALUE}.'))
-        value_token = tokens_for_value.consume()
-        if not tokens_for_value.is_null:
-            raise SingleInstructionInvalidArgumentException(_format('Superfluous arguments.'))
-        value_sdv = parse_string.parse_string_sdv_from_token(value_token,
-                                                             is_any_data_type())
-        executor = _SetExecutor(variable_name, value_sdv)
-        return TheInstructionEmbryo(executor, value_sdv.references)
+        var_name = token_parser.consume_mandatory_unquoted_string__w_err_msg(
+            False,
+            _MISSING_VAR_NAME_ERROR_MESSAGE
+        )
 
+        token_parser.report_superfluous_arguments_if_not_at_eol()
+        return TheInstructionEmbryo(_UnsetExecutor(var_name), [])
 
-UNSET_IDENTIFIER = 'unset'
-ASSIGNMENT_IDENTIFIER = instruction_arguments.ASSIGNMENT_OPERATOR
+    def _parse_set(self, token_parser: TokenParser, var_name: str) -> TheInstructionEmbryo:
+        token_parser.consume_mandatory_keyword(ASSIGNMENT_IDENTIFIER, False)
+
+        value = parse_string.parse_string_from_token_parser(token_parser, _VALUE_PARSE_CONFIGURATION)
+        token_parser.report_superfluous_arguments_if_not_at_eol()
+
+        executor = _SetExecutor(var_name, value)
+        return TheInstructionEmbryo(executor, value.references)
+
 
 PARTS_PARSER = PartsParserFromEmbryoParser(EmbryoParser(),
                                            MainStepResultTranslatorForUnconditionalSuccess())
@@ -195,13 +174,6 @@ class _UnsetExecutor(Executor):
             pass
 
 
-_MAIN_DESCRIPTION_REST_BODY = """\
-The manipulation affects all following phases.
-"""
-
-_ENV_VAR_REFERENCE = re.compile('\${[a-zA-Z0-9_]+}')
-
-
 def _expand_vars(value: str, environ: Dict[str, str]) -> str:
     def substitute(reference: str) -> str:
         var_name = reference[2:-1]
@@ -220,3 +192,47 @@ def _expand_vars(value: str, environ: Dict[str, str]) -> str:
         match = _ENV_VAR_REFERENCE.search(remaining)
     processed += remaining
     return processed
+
+
+class _MissingUnsetKeywordOrVarNameErrorMessage(token_stream_parser.ErrorMessageGenerator):
+    def message(self) -> str:
+        return _SF.format('Expecting {unset_keyword} or {var_name}')
+
+
+class _VarNameErrorMessage(token_stream_parser.ErrorMessageGenerator):
+    def message(self) -> str:
+        return _SF.format('Expecting {var_name}')
+
+
+_MISSING_UNSET_KEYWORD_OR_VAR_NAME_ERROR_MESSAGE = _MissingUnsetKeywordOrVarNameErrorMessage()
+_MISSING_VAR_NAME_ERROR_MESSAGE = _VarNameErrorMessage()
+
+_ENV_VAR_REFERENCE = re.compile('\${[a-zA-Z0-9_]+}')
+
+_VALUE_PARSE_CONFIGURATION = parse_string.Configuration(syntax_elements.STRING_SYNTAX_ELEMENT.singular_name)
+
+_VAR_NAME_ELEMENT = 'NAME'
+
+_SF = StringFormatter({
+    'unset_keyword': UNSET_IDENTIFIER,
+    'var_name': _VAR_NAME_ELEMENT,
+})
+
+_FORMAT_DICT = {
+    'NAME': _VAR_NAME_ELEMENT,
+    'VALUE': syntax_elements.STRING_SYNTAX_ELEMENT.singular_name,
+    'SYMBOLS': concepts.SYMBOL_CONCEPT_INFO.name.plural,
+    'unset_keyword': UNSET_IDENTIFIER,
+}
+
+_DESCRIPTION_OF_SET = """\
+Sets the environment variable {NAME} to {VALUE}.
+
+
+Elements of the form "${{var_name}}" in {VALUE}, will be replaced with the value of the environment variable "var_name",
+or the empty string, if there is no environment variable with that name.
+"""
+
+_MAIN_DESCRIPTION_REST_BODY = """\
+The manipulation affects all following phases.
+"""